+++ src/client/mod.rs
@@ -15,7 +15,7 @@ use crate::{
         raw::Raw,
     },
     config::{ClientConfig, MaximumPacketSize, ServerConfig, SessionExpiryInterval, SharedConfig},
-    fmt::{debug, error, panic, unreachable, warn},
+    fmt::{assert, debug, error, panic, unreachable, warn},
     header::{FixedHeader, PacketType},
     io::Transport,
     packet::{Packet, TxPacket},
@@ -55,6 +55,7 @@ pub use err::Error as MqttError;
 /// - `MAX_SUBSCRIPTION_IDENTIFIERS`: The maximum amount of subscription identifier properties the client can receive within a
 ///   single PUBLISH packet. If a packet with more subscription identifiers is received, the later identifers will be discarded.
 #[derive(Debug)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
 pub struct Client<
     'c,
     N: Transport,
@@ -166,12 +167,18 @@ impl<
         &self.session
     }
 
+    /// Returns an immutable reference to the supplied [`BufferProvider`] implementation.
+    #[inline]
+    pub fn buffer(&self) -> &B {
+        self.raw.buffer()
+    }
+
     /// Returns a mutable reference to the supplied [`BufferProvider`] implementation.
     ///
     /// This can for example be used to reset the underlying buffer if using `BumpBuffer`.
     #[inline]
-    pub fn buffer(&mut self) -> &mut B {
-        self.raw.buffer()
+    pub fn buffer_mut(&mut self) -> &mut B {
+        self.raw.buffer_mut()
     }
 
     /// Generates a new packet identifier.
@@ -285,7 +292,8 @@ impl<
                 options.keep_alive,
                 options.maximum_packet_size,
                 options.session_expiry_interval,
-                // Safety: `Self::new` panics if `RECEIVE_MAXIMUM` is 0 making this code unreachable.
+                // Safety: `Self::new` panics if `RECEIVE_MAXIMUM` is 0. Thus, this
+                // code is only reached when `RECEIVE_MAXIMUM` is greater than 0.
                 unsafe { NonZero::new_unchecked(RECEIVE_MAXIMUM as u16) },
                 options.request_response_information,
             );
@@ -1018,23 +1026,18 @@ impl<
                             reason_code,
                         })
                     }
-                    Some(s) => {
+                    Some(
+                        s @ CPublishFlightState::AwaitingPubrec
+                        | s @ CPublishFlightState::AwaitingPubcomp,
+                    ) => {
                         warn!("packet identifier {} in PUBACK is actually {:?}", pid, s);
 
                         // Readd this packet identifier to the session so that it can be republished
                         // after reconnecting.
 
-                        match s {
-                            CPublishFlightState::AwaitingPuback => unreachable!(),
-                            CPublishFlightState::AwaitingPubrec =>
-                            // Safety: Session::remove_cpublish returning Some and therefore successfully
-                            // removing a cpublish frees space to add a new in flight entry.
-                            unsafe { self.session.await_puback(pid) },
-                            CPublishFlightState::AwaitingPubcomp =>
-                            // Safety: Session::remove_cpublish returning Some and therefore successfully
-                            // removing a cpublish frees space to add a new in flight entry.
-                            unsafe { self.session.await_pubcomp(pid) },
-                        }
+                        // Safety: Session::remove_cpublish returning Some and therefore successfully
+                        // removing a cpublish frees space to add a new in flight entry.
+                        unsafe { self.session.r#await(pid, s) };
 
                         self.raw.close_with(Some(ReasonCode::ProtocolError));
                         return Err(MqttError::Server);
@@ -1084,23 +1087,18 @@ impl<
                             reason_code,
                         })
                     }
-                    Some(s) => {
+                    Some(
+                        s @ CPublishFlightState::AwaitingPuback
+                        | s @ CPublishFlightState::AwaitingPubcomp,
+                    ) => {
                         warn!("packet identifier {} in PUBREC is actually {:?}", pid, s);
 
                         // Readd this packet identifier to the session so that it can be republished
                         // after reconnecting.
 
-                        match s {
-                            CPublishFlightState::AwaitingPuback =>
-                            // Safety: Session::remove_cpublish returning Some and therefore successfully
-                            // removing a cpublish frees space to add a new in flight entry.
-                            unsafe { self.session.await_puback(pid) },
-                            CPublishFlightState::AwaitingPubrec => unreachable!(),
-                            CPublishFlightState::AwaitingPubcomp =>
-                            // Safety: Session::remove_cpublish returning Some and therefore successfully
-                            // removing a cpublish frees space to add a new in flight entry.
-                            unsafe { self.session.await_pubcomp(pid) },
-                        }
+                        // Safety: Session::remove_cpublish returning Some and therefore successfully
+                        // removing a cpublish frees space to add a new in flight entry.
+                        unsafe { self.session.r#await(pid, s) };
 
                         self.raw.close_with(Some(ReasonCode::ProtocolError));
                         return Err(MqttError::Server);
@@ -1193,23 +1191,18 @@ impl<
                             reason_code,
                         })
                     }
-                    Some(s) => {
+                    Some(
+                        s @ CPublishFlightState::AwaitingPuback
+                        | s @ CPublishFlightState::AwaitingPubrec,
+                    ) => {
                         warn!("packet identifier {} in PUBCOMP is actually {:?}", pid, s);
 
                         // Readd this packet identifier to the session so that it can be republished
                         // after reconnecting.
 
-                        match s {
-                            CPublishFlightState::AwaitingPuback =>
-                            // Safety: Session::remove_cpublish returning Some and therefore successfully
-                            // removing a cpublish frees space to add a new in flight entry.
-                            unsafe { self.session.await_puback(pid) },
-                            CPublishFlightState::AwaitingPubrec =>
-                            // Safety: Session::remove_cpublish returning Some and therefore successfully
-                            // removing a cpublish frees space to add a new in flight entry.
-                            unsafe { self.session.await_pubrec(pid) },
-                            CPublishFlightState::AwaitingPubcomp => unreachable!(),
-                        }
+                        // Safety: Session::remove_cpublish returning Some and therefore successfully
+                        // removing a cpublish frees space to add a new in flight entry.
+                        unsafe { self.session.r#await(pid, s) };
 
                         self.raw.close_with(Some(ReasonCode::ProtocolError));
                         return Err(MqttError::Server);