+++ src/fmt.rs
@@ -1,6 +1,27 @@
 #![allow(unused)]
 
 #[clippy::format_args]
+macro_rules! assert_ {
+    ($($x:tt)*) => {
+        {
+            #[cfg(not(feature = "defmt"))]
+            ::core::assert!($($x)*);
+            #[cfg(feature = "defmt")]
+            ::defmt::assert!($($x)*);
+        }
+    };
+}
+
+#[clippy::format_args]
+macro_rules! const_debug_assert_ {
+    ($($x:tt)*) => {
+        {
+            ::core::debug_assert!($($x)*);
+        }
+    };
+}
+
+#[clippy::format_args]
 macro_rules! debug_assert_ {
     ($($x:tt)*) => {
         {
@@ -68,7 +89,7 @@ macro_rules! trace {
             ::log::trace!($s $(, $x)*);
             #[cfg(feature = "defmt")]
             ::defmt::trace!($s $(, $x)*);
-            #[cfg(not(any(feature = "log", feature="defmt")))]
+            #[cfg(not(any(feature = "log", feature = "defmt")))]
             let _ = ($( & $x ),*);
         }
     };
@@ -82,7 +103,7 @@ macro_rules! debug {
             ::log::debug!($s $(, $x)*);
             #[cfg(feature = "defmt")]
             ::defmt::debug!($s $(, $x)*);
-            #[cfg(not(any(feature = "log", feature="defmt")))]
+            #[cfg(not(any(feature = "log", feature = "defmt")))]
             let _ = ($( & $x ),*);
         }
     };
@@ -96,7 +117,7 @@ macro_rules! info {
             ::defmt::info!($s $(, $x)*);
             #[cfg(feature = "log")]
             ::log::info!($s $(, $x)*);
-            #[cfg(not(any(feature = "log", feature="defmt")))]
+            #[cfg(not(any(feature = "log", feature = "defmt")))]
             let _ = ($( & $x ),*);
         }
     };
@@ -110,7 +131,7 @@ macro_rules! warn_ {
             ::log::warn!($s $(, $x)*);
             #[cfg(feature = "defmt")]
             ::defmt::warn!($s $(, $x)*);
-            #[cfg(not(any(feature = "log", feature="defmt")))]
+            #[cfg(not(any(feature = "log", feature = "defmt")))]
             let _ = ($( & $x ),*);
         }
     };
@@ -124,7 +145,7 @@ macro_rules! error {
             ::log::error!($s $(, $x)*);
             #[cfg(feature = "defmt")]
             ::defmt::error!($s $(, $x)*);
-            #[cfg(not(any(feature = "log", feature="defmt")))]
+            #[cfg(not(any(feature = "log", feature = "defmt")))]
             let _ = ($( & $x ),*);
         }
     };
@@ -135,6 +156,10 @@ pub(crate) use error;
 pub(crate) use info;
 pub(crate) use trace;
 pub(crate) use warn_ as warn;
+
+pub(crate) use assert_ as assert;
+
+pub(crate) use const_debug_assert_ as const_debug_assert;
 
 pub(crate) use debug_assert_ as debug_assert;
 pub(crate) use debug_assert_eq_ as debug_assert_eq;